move exposing metrics to proxy, instead of reg-service
