+++ pkg/proxy/proxy.go
@@ -54,6 +54,7 @@ type Proxy struct {
 	cl          client.Client
 	tokenParser *auth.TokenParser
 	spaceLister *handlers.SpaceLister
+	metrics     *handlers.Metrics
 }
 
 func NewProxy(app application.Application) (*Proxy, error) {
@@ -76,12 +77,13 @@ func newProxyWithClusterClient(app application.Application, cln client.Client) (
 
 	// init handlers
 	spaceLister := handlers.NewSpaceLister(app)
-
+	metrics := handlers.NewMetrics()
 	return &Proxy{
 		app:         app,
 		cl:          cln,
 		tokenParser: tokenParser,
 		spaceLister: spaceLister,
+		metrics:     metrics,
 	}, nil
 }
 
@@ -129,8 +131,8 @@ func (p *Proxy) StartProxy() *http.Server {
 	wg := router.Group("/apis/toolchain.dev.openshift.com/v1alpha1/workspaces")
 	wg.GET("/:workspace", p.spaceLister.HandleSpaceListRequest)
 	wg.GET("", p.spaceLister.HandleSpaceListRequest)
-
 	router.GET(proxyHealthEndpoint, p.health)
+	router.GET("/metrics", p.metrics.PrometheusHandler)
 	router.Any("/*", p.handleRequestAndRedirect)
 
 	// Insert the CORS preflight middleware
