+++ pkg/server/routes.go
@@ -26,7 +26,7 @@ func (srv *RegistrationServer) SetupRoutes() error {
 		authConfigCtrl := controller.NewAuthConfig()
 		analyticsCtrl := controller.NewAnalytics()
 		signupCtrl := controller.NewSignup(srv.application)
-		metricsCtrl := controller.NewMetrics()
+		//metricsCtrl := controller.NewMetrics()
 
 		// create the auth middleware
 		var authMiddleware *middleware.JWTMiddleware
@@ -41,7 +41,7 @@ func (srv *RegistrationServer) SetupRoutes() error {
 		unsecuredV1.GET("/health", healthCheckCtrl.GetHandler)
 		unsecuredV1.GET("/authconfig", authConfigCtrl.GetHandler)
 		unsecuredV1.GET("/segment-write-key", analyticsCtrl.GetDevSpacesSegmentWriteKey) //expose the devspaces segment key
-		unsecuredV1.GET("/metrics", metricsCtrl.PrometheusHandler)
+		//unsecuredV1.GET("/metrics", metricsCtrl.PrometheusHandler)
 		// secured routes
 		securedV1 := srv.router.Group("/api/v1")
 		securedV1.Use(authMiddleware.HandlerFunc())
